fix(eject,preset): make zudo-doc bin tsx-free + clarify packageOwnedRoutes default doc
Codex review (P2 x2) on base/stub-deletion-fast-follow:

1. eject bin required tsx, which a default generated project may not install
   (tsx is only added with the tags feature). The eject CLI needs no project TS
   config — only file copy/rewrite — so rewrite bin/zudo-doc.mjs as self-contained
   ESM that imports the COMPILED ../dist/eject + minimist/picocolors (declared
   package deps), with no tsx spawn. Delete the now-unused .ts runner. `zudo-doc
   eject` now runs on plain node in any generated project (#2367).

2. The packageOwnedRoutes JSDoc said "Default true" while buildPlugins reads the
   value truthily with no central defaulting (the brief's deliberate design — the
   default is realized via the value emitted into each consumer's settings.ts).
   Clarify the preset.ts + settings.ts JSDoc so "on by default" is not misread as
   a buildPlugins default; an omitted field is treated as off.
